MOVE namespace creation to its individual file
- Create namespace.go as part of operator internal package
- Move namespace functions there
- Renamed bundles.go to bundle.go to keep up it similar to the other names
